+++ installer/src/main/java/com/teamscale/profiler/installer/steps/InstallEtcEnvironmentStep.java
@@ -1,11 +1,10 @@
 package com.teamscale.profiler.installer.steps;
 
-import com.teamscale.profiler.installer.EnvironmentMap;
+import com.teamscale.profiler.installer.JvmEnvironmentMap;
 import com.teamscale.profiler.installer.FatalInstallerError;
 import com.teamscale.profiler.installer.PermissionError;
 import com.teamscale.profiler.installer.TeamscaleCredentials;
-import org.conqat.lib.commons.string.StringUtils;
-import org.conqat.lib.commons.system.SystemUtils;
+import org.apache.commons.lang3.SystemUtils;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -21,33 +20,35 @@ import java.util.stream.Collectors;
 public class InstallEtcEnvironmentStep implements IStep {
 
 	private final Path etcDirectory;
-	private final EnvironmentMap environmentVariables;
+	private final JvmEnvironmentMap environmentVariables;
 
-	public InstallEtcEnvironmentStep(Path etcDirectory, EnvironmentMap environmentMap) {
+	public InstallEtcEnvironmentStep(Path etcDirectory, JvmEnvironmentMap environmentMap) {
 		this.etcDirectory = etcDirectory;
 		this.environmentVariables = environmentMap;
 	}
 
 	@Override
-	public void install(TeamscaleCredentials credentials) throws FatalInstallerError {
-		if (!SystemUtils.isLinux()) {
-			return;
-		}
+	public boolean shouldRun() {
+		return SystemUtils.IS_OS_LINUX;
+	}
 
+	@Override
+	public void install(TeamscaleCredentials credentials) throws FatalInstallerError {
 		Path environmentFile = getEnvironmentFile();
+		String etcEnvironmentAddition = String.join("\n", environmentVariables.getEtcEnvironmentLinesList());
+
 		if (!Files.exists(environmentFile)) {
 			System.err.println(
 					environmentFile + " does not exist. Skipping system-wide registration of the profiler."
-							+ "\nYou need to manually register the profiler for process that should be profiled by"
-							+ " setting the following environment variables:"
-							+ "\n\n" + environmentVariables.getEtcEnvironmentString() + "\n");
+					+ "\nYou need to manually register the profiler for process that should be profiled by"
+					+ " setting the following environment variables:"
+					+ "\n\n" + etcEnvironmentAddition + "\n");
 			return;
 		}
 
-		String content = "\n" + environmentVariables.getEtcEnvironmentString() + "\n";
-
+		String content = "\n" + etcEnvironmentAddition + "\n";
 		try {
-			Files.write(environmentFile, content.getBytes(StandardCharsets.US_ASCII),
+			Files.writeString(environmentFile, content, StandardCharsets.US_ASCII,
 					StandardOpenOption.APPEND);
 		} catch (IOException e) {
 			throw new PermissionError("Could not change contents of " + environmentFile, e);
@@ -60,10 +61,6 @@ public class InstallEtcEnvironmentStep implements IStep {
 
 	@Override
 	public void uninstall(IUninstallErrorReporter errorReporter) {
-		if (!SystemUtils.isLinux()) {
-			return;
-		}
-
 		Path environmentFile = getEnvironmentFile();
 		if (!Files.exists(environmentFile)) {
 			return;
@@ -72,17 +69,16 @@ public class InstallEtcEnvironmentStep implements IStep {
 		try {
 			List<String> lines = Files.readAllLines(environmentFile, StandardCharsets.US_ASCII);
 			String newContent = removeProfilerVariables(lines);
-			Files.write(environmentFile, newContent.getBytes(StandardCharsets.US_ASCII));
+			Files.writeString(environmentFile, newContent, StandardCharsets.US_ASCII);
 		} catch (IOException e) {
 			errorReporter.report(new PermissionError("Failed to remove profiler from " + environmentFile + "." +
-					" Please remove the relevant environment variables yourself." +
-					" Otherwise, Java applications may crash.", e));
+													 " Please remove the relevant environment variables yourself." +
+													 " Otherwise, Java applications may crash.", e));
 		}
 	}
 
 	private String removeProfilerVariables(List<String> linesWithoutNewline) {
-		Set<String> linesToRemove = new HashSet<>(StringUtils.splitLinesAsList(
-				environmentVariables.getEtcEnvironmentString(), false));
+		Set<String> linesToRemove = new HashSet<>(environmentVariables.getEtcEnvironmentLinesList());
 		return linesWithoutNewline.stream().filter(line -> !linesToRemove.contains(line))
 				.collect(Collectors.joining("\n"));
 	}