+++ system-tests/tia-maven/src/test/java/com/teamscale/tia/TiaMavenMultipleJobsTest.java
@@ -1,11 +1,10 @@
 package com.teamscale.tia;
 
+import com.teamscale.test.commons.SystemTestUtils;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
-import org.conqat.lib.commons.io.ProcessUtils;
-import org.conqat.lib.commons.system.SystemUtils;
 import org.junit.jupiter.api.Test;
 
-import java.io.File;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -16,32 +15,27 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class TiaMavenMultipleJobsTest {
 
 	/**
-	 * Starts multiple maven processes and checks that the ports are dynamically set and the servers are correctly
+	 * Starts multiple Maven processes and checks that the ports are dynamically set and the servers are correctly
 	 * started.
 	 */
 	@Test
 	public void testMavenTia() throws Exception {
-		File workingDirectory = new File("maven-dump-local-project");
-		String executable = "./mvnw";
-		if (SystemUtils.isWindows()) {
-			executable = Paths.get("maven-dump-local-project", "mvnw.cmd").toUri().getPath();
-		}
+		String workingDirectory = "maven-dump-local-project";
+
 		// Clean once before testing parallel execution and make sure that the cleaning
 		// process is finished before testing.
-		ProcessUtils.ExecutionResult result = ProcessUtils
-				.execute(new ProcessBuilder(executable, "clean").directory(workingDirectory));
-		assertThat(result.terminatedByTimeoutOrInterruption()).isFalse();
+		SystemTestUtils.runMaven(workingDirectory, "clean");
+
+		// run three verify processes in parallel without waiting
 		for (int i = 0; i < 3; i++) {
-			new ProcessBuilder(executable, "verify").directory(workingDirectory).start();
+			SystemTestUtils.buildMavenProcess(workingDirectory, "verify").start();
 		}
-		result = ProcessUtils
-				.execute(new ProcessBuilder(executable, "verify").directory(workingDirectory));
-		// Get additional output for error debugging.
-		System.out.println(result.getStdout());
-		File configFile = new File(Paths.get(workingDirectory.getAbsolutePath(), "target", "tia", "agent.log").toUri());
-		String configContent = FileSystemUtils.readFile(configFile);
-		assertThat(configContent).isNotEmpty();
-		assertThat(result.terminatedByTimeoutOrInterruption()).isFalse();
-		assertThat(configContent).doesNotContain("Could not start http server on port");
+
+		// and one more that we wait for to terminate
+		SystemTestUtils.runMaven(workingDirectory, "verify");
+
+		Path configFile = Paths.get(workingDirectory, "target", "tia", "agent.log");
+		String configContent = FileSystemUtils.readFile(configFile.toFile());
+		assertThat(configContent).isNotEmpty().doesNotContain("Could not start http server on port");
 	}
 }